requirements.txt
streamlit==1.33.0
pandas
numpy
seaborn
Pillow
matplotlib
scikit-learn
